Rethrow an exception with stack trace information.
When a service throws one of the following RuntimeExceptions:
 * SecurityException
 * BadParcelableException
 * IllegalArgumentException
 * NullPointerException
 * IllegalStateException
It can be difficult to tell where the Exception occurs
from the client side. The Parcel only conveys the type
of Exception but does not include the stack trace information.

This change modifies the Parcel to convey the stack trace information
when it is an eng build or a userdebug build.
The 'cause' section in the client side exception contains
the server side exception with the stack trace information.
When the client side exception is output, the 'cause' section
is also output at the same time.

Change-Id: I9449b9ea02825361dd51b57ad250b9d45d913aef